+++ jest.config.cjs
@@ -11,4 +11,5 @@ module.exports = {
   transform: {
     '\\.(ts)$': 'ts-jest',
   },
+  verbose: true
 }
