+++ models/label.ts
@@ -2,7 +2,7 @@ import * as t from 'io-ts'
 
 import { ImageIO, ResourceURLIO } from './commons'
 import { makeEnumIOType } from './helpers'
-import { DataQualityEnum } from '../src/constants'
+import { DataQualityEnum } from '../src/enums'
 
 /**
  * @internal
