+++ models/release.ts
@@ -2,7 +2,7 @@ import * as t from 'io-ts'
 
 import { ImageIO, ResourceURLIO, StatNumberIO, VideoIO } from './commons'
 import { makeEnumIOType } from './helpers'
-import { CommunityStatusesEnum, DataQualityEnum } from '../src/constants'
+import { CommunityStatusesEnum, DataQualityEnum } from '../src/enums'
 
 /**
  * @internal
