+++ src/constants.ts
@@ -1,0 +1,27 @@
+/**
+ * Base API URL to which URI will be appended.
+ *
+ * @internal
+ */
+export const API_BASE_URL = 'https://api.discogs.com'
+
+/**
+ * Base URL dedicated to Discogs images.
+ *
+ * @internal
+ */
+export const IMG_BASE_URL = 'https://img.discogs.com'
+
+/**
+ * Discogs API version.
+ *
+ * @internal
+ */
+export const API_VERSION = 'v2'
+
+/**
+ * Default user-agent to be used in requests.
+ *
+ * @internal
+ */
+export const DEFAULT_USER_AGENT = `Discojs/__packageVersion__`
