+++ src/enums/conditions.ts
