+++ src/enums/currencies.ts
