+++ src/enums/inventory.ts
