+++ src/enums/searchTypes.ts
