+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './constants'
 export * from './discojs'
+export * from './enums'
 export * from './errors'
 export { SortOrdersEnum } from './utils'
