+++ src/inventoryExport.ts
@@ -1,0 +1,3 @@
+// @TODO: Introduce methods.
+
+export class InventoryExport {}
