+++ src/userLists.spec.ts
@@ -1,0 +1,37 @@
+import * as t from 'io-ts'
+
+import { UserListItemsResponseIO, UserListsResponseIO } from '../models'
+import { Discojs } from './discojs'
+
+declare const client: Discojs
+
+const rodneyfool = 'rodneyfool'
+const pagination = { page: 1, perPage: 1 }
+
+describe('User Lists', () => {
+  describe('User Lists', () => {
+    it('getListsForUser', async () => {
+      const apiResponse = await client.getListsForUser(rodneyfool)
+      expect(t.exact(UserListsResponseIO).is(apiResponse)).toBeTruthy()
+    })
+
+    it('getListsForUser - with pagination', async () => {
+      const apiResponse = await client.getListsForUser(rodneyfool, pagination)
+      expect(apiResponse.pagination).toHaveProperty('page', pagination.page)
+      expect(apiResponse.pagination).toHaveProperty('per_page', pagination.perPage)
+    })
+
+    it('getLists', async () => {
+      const apiResponse = await client.getLists()
+      expect(t.exact(UserListsResponseIO).is(apiResponse)).toBeTruthy()
+    })
+  })
+
+  describe('List', () => {
+    it('getListItems', async () => {
+      const listId = 962557
+      const apiResponse = await client.getListItems(listId)
+      expect(t.exact(UserListItemsResponseIO).is(apiResponse)).toBeTruthy()
+    })
+  })
+})
