+++ src/utils/applyMixins.ts
@@ -1,0 +1,12 @@
+// https://www.typescriptlang.org/docs/handbook/mixins.html
+export function applyMixins(derivedCtor: any, constructors: any[]) {
+  constructors.forEach((baseCtor) => {
+    Object.getOwnPropertyNames(baseCtor.prototype).forEach((name) => {
+      Object.defineProperty(
+        derivedCtor.prototype,
+        name,
+        Object.getOwnPropertyDescriptor(baseCtor.prototype, name) || Object.create(null),
+      )
+    })
+  })
+}
